fix(sysadmin): cast windowMonths to int for make_interval (#914)
Prisma 6's $queryRaw binds JS number parameters as int8, which makes
`make_interval(months => $1 - 1)` resolve to `make_interval(bigint)` and
fail with PostgreSQL error 42883. Cast the parameter to int so the
integer-typed overload is selected.

Fixes 500s on sysAdminCommunityDetail / sysAdminDashboard.

+++ src/application/domain/sysadmin/data/repository.ts
@@ -245,7 +245,7 @@ export default class SysAdminRepository implements ISysAdminRepository {
             DATE_TRUNC(
               'month',
               (${asOf}::timestamp AT TIME ZONE 'UTC' AT TIME ZONE 'Asia/Tokyo')
-            ) - make_interval(months => ${windowMonths} - 1),
+            ) - make_interval(months => (${windowMonths})::int - 1),
             DATE_TRUNC(
               'month',
               (${asOf}::timestamp AT TIME ZONE 'UTC' AT TIME ZONE 'Asia/Tokyo')
